refactor(workstations): preserve package layout
